% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vctr.R
\name{card_suit}
\alias{card_suit}
\title{Generic to get card suit (e.g., "red", "blue", "wild")}
\usage{
card_suit(x)
}
\arguments{
\item{x}{A card_vctr}
}
\value{
Character vector of suits
}
\description{
Generic to get card suit (e.g., "red", "blue", "wild")
}
