% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vctr.R
\name{card_value}
\alias{card_value}
\title{Generic to get card value (e.g., "3", "skip", "draw4")}
\usage{
card_value(x)
}
\arguments{
\item{x}{A card_vctr}
}
\value{
Character vector of values
}
\description{
Generic to get card value (e.g., "3", "skip", "draw4")
}
