% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vctr.R
\name{convert_to_card_vctr}
\alias{convert_to_card_vctr}
\title{Convert a tibble to a card_vctr}
\usage{
convert_to_card_vctr(tbl)
}
\arguments{
\item{tbl}{A tibble with columns `color` and `value`}
}
\value{
A `card_vctr` vector
}
\description{
Convert a tibble to a card_vctr
}
