% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uno_deck.R
\name{create_uno_deck}
\alias{create_uno_deck}
\title{Creating a UNO deck}
\usage{
create_uno_deck()
}
\value{
A tibble with columns: color, value, type.
}
\description{
Builds a standard 108-card UNO deck with numbers, actions, and wild cards.
}
