% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vctr.R
\name{new_card_vctr}
\alias{new_card_vctr}
\title{Create a card vector (custom vctrs class)}
\usage{
new_card_vctr(x = character())
}
\arguments{
\item{x}{Character vector like "red_3", "blue_skip", etc.}
}
\value{
A custom `card_vctr` object
}
\description{
Creates a new UNO card vector using a custom vctrs class.
}
\examples{
new_card_vctr(c("red_3", "blue_skip", "wild_draw4"))
}
