% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_hands.R
\name{deal_hands}
\alias{deal_hands}
\title{Deal 7 cards to each player after shuffling the deck}
\usage{
deal_hands(deck, n_players = 4)
}
\arguments{
\item{deck}{A tibble with UNO cards created using create_deck()}

\item{n_players}{Number of players}
}
\value{
A list with:
hands - list of individual player card set,
deck - remaining cards forming the draw pile and
discard - the starting card for the game to begin
}
\description{
Deal 7 cards to each player after shuffling the deck
}
