% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play_game}
\alias{play_game}
\title{Game logic}
\usage{
play_game(n_players = 4)
}
\description{
The gmae logic is divided into teo main functions
}
\details{
`play_game()`-  Initializes the game by creating a deck, dealing hands to players, and starting the game loop.
 `play_turns_loop()` - Handles the main game loop, processing each player's turn until a winner is found.
 This included drawing cards, playing cards, and handling special actions like skips and reverses. Logic for
  wild cards and drawing cards is also included.
}
