% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_helpers.R
\name{summarise_hand}
\alias{summarise_hand}
\title{Summarise the hand of a specific player}
\usage{
summarise_hand(hands, player)
}
\arguments{
\item{hands}{A named list of `card_vctr` hands.}

\item{player}{A character string, e.g., "Player_1".}
}
\value{
A named integer vector with counts of each color.
}
\description{
Given a named list of player hands, this function summarizes how many cards
of each color a specific player has.
}
